refactor: LogBuffer to LogStep
