+++ modules/logger/.changes/000-little-mugs-vanish.md
@@ -1,0 +1,7 @@
+---
+type: minor
+---
+
+The `Logger` and `LogStep` now implement streaming output differently in order to always fully capture potential output and switch on verbosity.
+
+This is a major internal rewrite, but should be fully compatible with the previous implementations.
