+++ modules/logger/src/utils/string.ts
@@ -1,0 +1,30 @@
+export function stringify(item: unknown): string {
+	if (typeof item === 'string') {
+		return item.replace(/^\n+/, '').replace(/\n*$/g, '');
+	}
+
+	if (
+		Array.isArray(item) ||
+		(typeof item === 'object' && item !== null && item.constructor === Object) ||
+		item === null
+	) {
+		return JSON.stringify(item, null, 2);
+	}
+
+	if (item instanceof Date) {
+		return item.toISOString();
+	}
+
+	if (typeof item === 'function' && item.length === 0) {
+		return stringify(item());
+	}
+
+	return `${String(item)}`;
+}
+
+export function ensureNewline(str: string): string {
+	if (/^\S*$/.test(str)) {
+		return '';
+	}
+	return str.endsWith('\n') ? str : str.replace(/\n*$/g, '\n');
+}
