+++ modules/subprocess/src/index.ts
@@ -132,13 +132,11 @@ export async function run(options: RunSpec): Promise<[string, string]> {
 		${JSON.stringify(withoutLogger, null, 2)}\n`,
 			);
 
-			return Promise.resolve()
-				.then(() => {
-					return !inputStep ? step.end() : Promise.resolve();
-				})
-				.then(() => {
-					resolve([out.trim(), err.trim()]);
-				});
+			if (!inputStep) {
+				step.end();
+			}
+
+			return resolve([out.trim(), err.trim()]);
 		}
 
 		if (inputStep) {
@@ -161,14 +159,11 @@ ${JSON.stringify(withoutLogger, null, 2)}\n${process.env.ONEREPO_ROOT ?? process
 
 		subprocess.on('error', (error) => {
 			!options.skipFailures && step.error(error);
-			return Promise.resolve()
-				.then(() => {
-					logger.unpause();
-					return !inputStep ? step.end() : Promise.resolve();
-				})
-				.then(() => {
-					reject(error);
-				});
+			logger.unpause();
+			if (!inputStep) {
+				step.end();
+			}
+			return reject(error);
 		});
 
 		if (subprocess.stdout && subprocess.stderr) {
@@ -200,23 +195,24 @@ ${JSON.stringify(withoutLogger, null, 2)}\n${process.env.ONEREPO_ROOT ?? process
 				const error = new SubprocessError(`${sortedOut || code}`);
 				step.error(sortedOut.trim());
 				step.error(`Process exited with code ${code}`);
-				return (!inputStep ? step.end() : Promise.resolve()).then(() => {
-					logger.unpause();
-					reject(error);
-				});
+				if (!inputStep) {
+					step.end();
+				}
+
+				logger.unpause();
+				reject(error);
+				return;
 			}
 
 			if (inputStep) {
 				step.timing(`onerepo_start_Subprocess: ${options.name}`, `onerepo_end_Subprocess: ${options.name}`);
 			}
 
-			return Promise.resolve()
-				.then(() => {
-					return !inputStep ? step.end() : Promise.resolve();
-				})
-				.then(() => {
-					resolve([out.trim(), err.trim()]);
-				});
+			if (!inputStep) {
+				step.end();
+			}
+
+			return resolve([out.trim(), err.trim()]);
 		});
 	});
 }
@@ -309,12 +305,9 @@ export async function sudo(options: Omit<RunSpec, 'opts'> & { reason?: string })
 				logger.log(stdout);
 				logger.log(stderr);
 
-				return Promise.resolve()
-					.then(() => {
-						logger.unpause();
-						return step.end();
-					})
-					.then(() => resolve([stdout, stderr]));
+				logger.unpause();
+				step.end();
+				resolve([stdout, stderr]);
 			},
 		);
 	});
